requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.1
decorator==5.1.0
dj-database-url==0.5.0
Django==3.2.8
future==0.18.2
gunicorn==20.1.0
psycopg2-binary==2.9.2
pycparser==2.21
python-telegram-bot==12.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
